hämtaKonto() färdig

+++ databasAPI/IDatabas.java
@@ -28,7 +28,7 @@ public interface IDatabas {
      * behöver definiera vad databasen kan tänkas svara... för tester
      *
      * */
-    String läggTillSvartlista(BigInteger personNr);
+    String läggTillSvartlista(long personNr);
 
 
     /*returnera String att Konto är skapad*/
@@ -37,7 +37,7 @@ public interface IDatabas {
     * behöver definiera vad databasen kan tänkas svara... för tester
     *
     * */
-    String skapaKonto(String fnamn, String  enamn, BigInteger personNr, String roll);
+    String skapaKonto(String fnamn, String  enamn, long personNr, String roll);
 
 
     /* returnera en String efter att Konto är avslutad.
